Laurie Heath  laurieheath@btinternet.com
Added details buttons to main form.
 

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@23 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/Menu/Options.cs
@@ -23,8 +23,8 @@ namespace ORTS
             RegistryKey RK = Registry.CurrentUser.OpenSubKey(Program.RegistryKey);
             if (RK != null)
             {
-                this.numericWorldObjectDensity.Value = (int)RK.GetValue("WorldObjectDensity", numericWorldObjectDensity.Value);
-                this.numericSoundDetailLevel.Value = (int)RK.GetValue("SoundDetailLevel", numericSoundDetailLevel.Value);
+                this.numericWorldObjectDensity.Value = (int)RK.GetValue("WorldObjectDensity", (int)numericWorldObjectDensity.Value);
+                this.numericSoundDetailLevel.Value = (int)RK.GetValue("SoundDetailLevel", (int)numericSoundDetailLevel.Value);
 
             }
         }
